+++ methods/src/main/kotlin/methods/integerrationalexponents/IntegerRationalExponentsPlans.kt
@@ -9,6 +9,7 @@ import engine.patterns.IntegerFractionPattern
 import engine.patterns.UnsignedIntegerPattern
 import engine.patterns.powerOf
 import methods.fractionarithmetic.FractionArithmeticPlans
+import methods.fractionarithmetic.addIntegerAndFraction
 import methods.fractionarithmetic.addIntegerFractions
 import methods.general.GeneralRules
 import methods.general.NormalizationRules
@@ -82,6 +83,7 @@ enum class IntegerRationalExponentsPlans(override val runner: CompositeMethod) :
             steps {
                 apply(GeneralRules.RewriteProductOfPowersWithSameBase)
                 firstOf {
+                    option { deeply(addIntegerAndFraction) }
                     option { deeply(addIntegerFractions) }
                     option { deeply(IntegerArithmeticRules.EvaluateSignedIntegerAddition) }
                 }
@@ -193,6 +195,7 @@ val simplifyRationalExponentsInProduct = steps {
             option { deeply(IntegerRationalExponentsPlans.SimplifyFractionOfPowersWithSameBase) }
             option { deeply(IntegerRationalExponentsPlans.SimplifyFractionOfPowersWithSameExponent) }
             option { deeply(IntegerRationalExponentsPlans.ApplyPowerRuleOfExponents) }
+            option { deeply(IntegerRationalExponentsRules.ApplyReciprocalPowerRule) }
             option { deeply(IntegerRationalExponentsPlans.SimplifyProductOfPowersWithRationalExponents) }
         }
     }